requirements.txt
python-docx>=1.1.0
qdrant-client>=1.9.0
requests>=2.31.0
fastapi>=0.115.0
uvicorn>=0.30.0
